Refactor Parser
This was done via Extract Function, to make it such that the main
Parser class looks cleaner, with just getter functions to get the
appropriate UserCommand object for each command.

Each getter function appropriately handles the logic and exception
handling for the Command it is associated with.

JavaDocs for all newly created methods are added as well.